+++ dungeon-game-core/src/com/dungeon/game/item/Slot.java
@@ -65,7 +65,7 @@ public class Slot {
 		if(((world.mouse.slot.item==null?true:world.mouse.slot.item.type==type)||type==0)) {
 			if(world.mouse.shift_down) {
 				if(world.mouse.rb_pressed  && item != null) {
-					Window temp = new DescWindow(world, (int)world.mouse.x, (int)world.mouse.y, item);
+					Window temp = new DescWindow(world, world.mouse.x, world.mouse.y, item);
 					temp.open();
 				}
 			}
